clean up - reorg rle
